fix: use fs.stat to detect directory index
`fs.ls` tells you the type of every directory entry - to do this it
loads the first block of each entry to parse the unixfs metadata -
this is obviously a bad idea for very large directories.

Switch to `fs.stat` which only loads the directory entry we are
interested in.