fix: silence max listeners exceeded warning
Each request can kick of a lot of abortable work so the abort signal
created will have a lot of listeners attached to it.

This is expected, but node will print a warning message that people
can mistake for an error so increase the number of listeners allowed
to silence the warning.