fix: remove uncaught exception handler
There are no circumstances under which it is safe to ignore uncaught
exceptions, since the application is left in an unconsistent state.

The node documentation is unequivocal about this:

https://nodejs.org/api/process.html#warning-using-uncaughtexception-correctly

Any code that can throw an exception should be wrapped in a `try/catch`
or there should be some other method of handling the error.